refactor: move index helper bodies into chunk_index.c (Phase 3b)
Follow-up tidiness to the Phase 3 ChunkIndex extraction. Moves the
static helper bodies that operate on ChunkIndex state out of
chunk_store.c and into chunk_index.c:

  - csReadIndex (mmap-or-malloc-read + deserialize the durable index)
  - csSearchIndex (binary search by hash)
  - csMergeIndex (merge pending+index into a new sorted array)
  - csReleaseIndexBuf (unmap or free)
  - csDeserializeIndexEntry, csIndexEntryCmp, csIndexLowerBound,
    csMapIndex (3 platform variants), csUnmapIndex (3 variants)

The four functions chunk_store.c still calls (csReadIndex,
csSearchIndex, csMergeIndex, csReleaseIndexBuf) become non-static
and get forward declarations in chunk_index.h. The rest stay
static-internal to chunk_index.c.

chunk_index.c now needs `#include "chunk_store.h"` for the
ChunkStore struct (used by csReadIndex and csMergeIndex which read
cs->file.* and cs->staging.*). The CS_READ_U32 / CS_READ_I64 byte
unpacking macros are duplicated locally to avoid pulling more into
chunk_index.h.

No behavior change. No on-disk format change.

Build clean. Smoke (CREATE TABLE + dolt_commit) returns hash.
lint_layers passes. Index-touching regressions all pass:
wal_offset_corruption (5/5), truncated_wal_rejected (6/6),
memory_chunk_lookup_corruption (5/5).

chunk_store.c is now ~336 lines smaller. The remaining
Phase 2b/4b/5b/1b-equivalent function moves (refs, staging, wal, file)
remain for future PRs.
